refactor(hybridconv): park the dual/triple convolver API as dormant code
Audit #146 TD009, maintainer decision 2026-07-04: the low-latency
dual/triple-segment API has no callers but is reserved for planned
low-latency (and possibly multicore) work on heavy convolution stacks,
so instead of lingering as ambient dead weight inside the live TU it
moves to libHybridConv_eapo_dormant.{h,cpp} - compiled by NO project,
with a banner stating why it exists and how to revive it. The live
files shrink by ~290 lines and keep pointer comments.

Verified: the dormant TU still compiles standalone against the FFTW
headers (revivability proof), and Common + all four test suites pass
unchanged (MultiConvolution 3421 checks).
